requirements.txt
djangorestframework==3.4.4
markdown==2.6.6
django-filter==0.14.0
Django==1.8
django-jsonfield==1.0.1
pygments==2.1.3